feat: reorganize codeanalysis project for combined featuresets
