+++ src/Daybreak.CodeAnalysis/Common/AbstractCodeFixProvider.cs
