+++ src/Daybreak.CodeAnalysis/Common/AbstractDiagnosticAnalyzer.cs
