+++ src/Daybreak.CodeAnalysis/Hooks/AttributeHookPair.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
 
-namespace Daybreak.CodeAnalysis;
+namespace Daybreak.CodeAnalysis.Hooks;
 
 public sealed record AttributeHookPair(
     INamedTypeSymbol AttributeClass,
